refactor: implement MVVM architecture with modular components
- Extract FileExplorerPanel from ContentView (790+ lines removed)
- Extract InstructionsPanel from ContentView for better separation
- Create ContentViewViewModel following MVVM pattern
- Add FileIconProvider to eliminate 100+ lines of code duplication
- Break monolithic ContentView into focused, reusable components
- Add ArrayExtensions utility for batch processing
- Improve code maintainability and testability

ContentView.swift: 1,347 → 557 lines (-59% reduction)
