+++ Sources/ArrayExtensions.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension Array {
+    /// Splits the array into chunks of the specified size
+    func chunked(into size: Int) -> [[Element]] {
+        return stride(from: 0, to: count, by: size).map {
+            Array(self[$0..<Swift.min($0 + size, count)])
+        }
+    }
+}
